Rename `peace_resources::dir` to `peace_resources::paths`.
