+++ crate/resources/src/lib.rs
@@ -8,8 +8,8 @@ pub use type_reg;
 
 pub use crate::{item_spec_rt_id::ItemSpecRtId, resources::Resources};
 
-pub mod dir;
 pub mod internal;
+pub mod paths;
 pub mod resources_type_state;
 pub mod states;
