+++ crate/resources/src/paths.rs
